Expanded/Maze/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.1
networkx==3.2.1
numpy==1.23.5
opencv-python==4.8.0.74
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
tifffile==2024.7.24
urllib3==2.2.2
zipp==3.19.2
